+++ tests/modules_test/math/geom.scm
@@ -1,0 +1,3 @@
+(define (square x) (* x x))
+(define (hypot a b)
+  (sqrt (+ (square a) (square b))))