+++ tests/modules_test/pkg/mod.scm
@@ -1,0 +1,2 @@
+(define (greet name)
+  (format "Hello from pkg, {}!" name))